garage-voice-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "garage-voice-agent"
version = "0.1.0"
description = "Agent vocal LiveKit pour garages et carrosseries en France"
requires-python = ">=3.11,<3.15"
dependencies = [
    "fastapi>=0.115,<1",
    "livekit-agents[deepgram,elevenlabs,openai,silero,turn-detector]~=1.5",
    "livekit-plugins-ai-coustics~=0.2",
    "pydantic>=2.8,<3",
    "pydantic-settings>=2.4,<3",
    "python-dotenv>=1.0,<2",
    "uvicorn[standard]>=0.32,<1",
]

[project.scripts]
garage-voice-dev = "dev_runner:main"
garage-voice-doctor = "doctor:main"

[dependency-groups]
dev = [
    "httpx>=0.27",
    "pytest>=8.2",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src", "."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
